Move file parse logic from FileParser to FileParseHelper
